pyproject.toml
[tool.poetry]
name = "lulu-database"
version = "0.1.0"
description = ""
authors = ["Simón <simon.garcial@udea.edu.co>"]
readme = "README.md"
packages = [{include = "lulu-database"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.5"}
tortoise-orm = "^0.19.3"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.20.0"
autopep8 = "^2.0.1"
black = "^23.1.0"
flake8 = "^6.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
